Check if file type is acceptable
I've added a simple check for whether the file ends in ".js" or ".ts" to avoid parsing errors when processing other file types. This allows us to store templates and other route related files alongside the route definition without the need to prepend a dot or underscore character to the filename.